[chore] #118 사용하지 않는 파일 제거 및 파일 분리
