+++ app/src/main/java/org/sopt/pingle/presentation/ui/main/plan/planlocation/PlanLocationAdapter.kt
@@ -6,12 +6,16 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.pingle.databinding.ItemPlanLocationBinding
 import org.sopt.pingle.domain.model.PlanLocationEntity
+import org.sopt.pingle.util.view.ItemDiffCallback
 
 class PlanLocationAdapter(
     private val setOldItem: (Int) -> Unit
 ) :
     ListAdapter<PlanLocationEntity, PlanLocationViewHolder>(
-        PlanLocationDiffCallback()
+        ItemDiffCallback<PlanLocationEntity>(
+            onItemsTheSame = { old, new -> old.location == new.location },
+            onContentsTheSame = { old, new -> old == new }
+        )
     ) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PlanLocationViewHolder =
@@ -22,18 +26,5 @@ class PlanLocationAdapter(
 
     override fun onBindViewHolder(holder: PlanLocationViewHolder, position: Int) {
         holder.onBind(getItem(position))
-    }
-}
-
-class PlanLocationViewHolder(
-    private val binding: ItemPlanLocationBinding,
-    private val setOldItem: (Int) -> Unit
-) : RecyclerView.ViewHolder(binding.root) {
-    fun onBind(item: PlanLocationEntity) {
-        binding.planLocationItem = item
-
-        binding.root.setOnClickListener {
-            setOldItem(bindingAdapterPosition)
-        }
     }
 }
